+++ NEWS.md
@@ -1,5 +1,8 @@
 # remotes (development version)
 
+* `install_git()` now can take credentials from the global option
+  `remotes.git_credentials` (#378).
+
 * `install_git()` now works with SHA references and external git (#389).
 
 * GitHub remotes that point to branches no longer fail when the branch is later